moved test template in other forlders
