+++ src/js/templates/index.mjs
@@ -1,1 +1,2 @@
 export * from "./post.mjs";
+export * from "./posts.mjs";
